DRTII-571 Move request classes to more sensible location
