+++ server/src/main/scala/actors/FlightsStateActor.scala
@@ -1,5 +1,6 @@
 package actors
 
+import actors.PartitionedPortStateActor.{DateRangeLike, GetFlights, GetFlightsForTerminalDateRange, GetStateForDateRange, GetStateForTerminalDateRange, GetUpdatesSince, PointInTimeQuery}
 import actors.PortStateMessageConversion._
 import actors.acking.AckingReceiver.{Ack, StreamCompleted, StreamFailure, StreamInitialized}
 import actors.daily.{RequestAndTerminate, RequestAndTerminateActor}
@@ -21,7 +22,6 @@ import scalapb.GeneratedMessage
 import server.protobuf.messages.CrunchState._
 import server.protobuf.messages.FlightsMessage.UniqueArrivalMessage
 import services.SDate
-import services.crunch.deskrecs.{GetFlightsForDateRange, GetStateForDateRange, GetStateForTerminalDateRange}
 
 import scala.concurrent.ExecutionContextExecutor
 import scala.concurrent.duration._
@@ -166,7 +166,7 @@ class FlightsStateActor(val now: () => SDateLike,
   }
 
   def toLegacyMessage(message: DateRangeLike): DateRangeLike = message match {
-    case GetStateForDateRange(from, to) => GetFlightsForDateRange(from, to)
+    case GetStateForDateRange(from, to) => GetFlights(from, to)
     case GetStateForTerminalDateRange(from, to, terminal) => GetFlightsForTerminalDateRange(from, to, terminal)
   }
