+++ server/src/main/scala/services/crunch/deskrecs/RunnableDeployments.scala
@@ -1,5 +1,6 @@
 package services.crunch.deskrecs
 
+import actors.PartitionedPortStateActor.GetStateForDateRange
 import actors.acking.AckingReceiver._
 import akka.actor.ActorRef
 import akka.pattern.ask
