+++ server/src/main/scala/services/imports/ArrivalImporter.scala
@@ -1,6 +1,7 @@
 package services.imports
 
 import actors.GetState
+import actors.PartitionedPortStateActor.GetFlights
 import actors.acking.AckingReceiver.{Ack, StreamInitialized}
 import akka.actor.{Actor, ActorLogging}
 import akka.pattern.pipe
@@ -9,10 +10,9 @@ import drt.shared.FlightsApi.FlightsWithSplits
 import drt.shared.PaxTypes._
 import drt.shared.SplitRatiosNs.SplitSources.ApiSplitsWithHistoricalEGateAndFTPercentages
 import drt.shared.Terminals.Terminal
-import drt.shared.api.Arrival
 import drt.shared._
+import drt.shared.api.Arrival
 import services.SDate
-import services.crunch.deskrecs.GetFlightsForDateRange
 
 import scala.concurrent.{ExecutionContextExecutor, Promise}
 import scala.util.Try
@@ -117,7 +117,7 @@ class ArrivalCrunchSimulationActor(fws: FlightsWithSplits) extends Actor with Ac
   var promisedResult: Promise[DeskRecMinutes] = Promise[DeskRecMinutes]
 
   override def receive: Receive = {
-    case GetFlightsForDateRange(_, _) =>
+    case GetFlights(_, _) =>
       sender() ! fws
 
     case GetState =>
