+++ server/src/test/scala/services/crunch/deskrecs/RunnableDeploymentsSpec.scala
@@ -1,6 +1,7 @@
 package services.crunch.deskrecs
 
 import actors.MinuteLookupsLike
+import actors.PartitionedPortStateActor.GetStateForDateRange
 import actors.acking.AckingReceiver.{Ack, StreamCompleted, StreamFailure, StreamInitialized}
 import actors.daily.RequestAndTerminateActor
 import actors.minutes.MinutesActorLike.{MinutesLookup, MinutesUpdate}
