+++ server/src/test/scala/services/crunch/deskrecs/RunnableDeskRecsSpec.scala
@@ -1,5 +1,6 @@
 package services.crunch.deskrecs
 
+import actors.PartitionedPortStateActor.GetFlights
 import actors.acking.AckingReceiver.{Ack, StreamCompleted, StreamFailure, StreamInitialized}
 import akka.actor.{Actor, Props}
 import akka.testkit.TestProbe
@@ -39,7 +40,7 @@ class MockPortStateActor(probe: TestProbe, responseDelayMillis: Long = 0L) exten
       log.error(s"Failed", t)
       probe.ref ! StreamFailure
 
-    case getFlights: GetFlightsForDateRange =>
+    case getFlights: GetFlights =>
       Thread.sleep(responseDelayMillis)
       sender() ! FlightsWithSplits(flightsToReturn.map(fws => (fws.unique, fws)))
       probe.ref ! getFlights
@@ -88,7 +89,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     val expectedEnd = midnight20190101.addMinutes(30).millisSinceEpoch
 
     portStateProbe.fishForMessage(1 second) {
-      case GetFlightsForDateRange(start, end) => start == expectedStart && end == expectedEnd
+      case GetFlights(start, end) => start == expectedStart && end == expectedEnd
     }
 
     success
@@ -109,7 +110,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     val expectedEnd = midnight20190101.addMinutes(30).millisSinceEpoch
 
     portStateProbe.fishForMessage(1 second) {
-      case GetFlightsForDateRange(start, end) => start == expectedStart && end == expectedEnd
+      case GetFlights(start, end) => start == expectedStart && end == expectedEnd
     }
     portStateProbe.fishForMessage(1 second) {
       case _: DeskRecMinutes => true
